Exercises 8.20(Adding items ) & 8.21(Explain)
explanation- item descriptions should be produced by the Item class to adhere to the principle of encapsulation, since the Item class contains all the data about the item. The responsibility of displaying these descriptions goes to the Room class, as it runs the environment's narrative, including the presence and portrayal of items within it. the Item class focuses on item specific information and representation, while the Room class manages the environment and how items are integrated into room descriptions.